conf: receive syslog on udp:1514
Allocate port 1514 in the docker configuration. A simple logstash
configuration is used. The process timestamp is added to the metadata to
debug potential timesync issues. udp is used for simplicity as the
intended environment is a closed network; if message loss is observed a
more robust protocol will be considered.